move tabs only used in docs
