+++ apps/docs/features/ui/Tabs.tsx
@@ -1,9 +1,11 @@
 'use client'
 
-import { useCallback, type ComponentPropsWithoutRef, type PropsWithChildren } from 'react'
-import { Tabs as TabsPrimitive, type TabsProps } from 'ui'
-import { withQueryParams, withSticky, type QueryParamsProps } from 'ui-patterns/ComplexTabs'
 import { useTocRerenderTrigger } from '~/features/docs/GuidesMdx.state'
+import { useCallback, type ComponentPropsWithoutRef, type PropsWithChildren } from 'react'
+
+import { Tabs as TabsPrimitive, type TabsProps } from './UITabs'
+import { withQueryParams, type QueryParamsProps } from './withQueryParams'
+import { withSticky } from './withSticky'
 
 const TabsWithStickyAndQueryParams = withSticky<PropsWithChildren<TabsProps & QueryParamsProps>>(
   withQueryParams(TabsPrimitive)
