+++ apps/docs/package.json
@@ -70,6 +70,7 @@
     "acorn-typescript": "^1.4.13",
     "ai-commands": "workspace:*",
     "beautiful-mermaid": "^1.1.3",
+    "class-variance-authority": "^0.7.1",
     "common": "workspace:*",
     "common-tags": "^1.8.2",
     "config": "workspace:*",
