--- packages/ui-patterns/src/ComplexTabs/index.tsx
@@ -1,2 +1,0 @@
-export * from './withQueryParams'
-export * from './withSticky'
