--- packages/ui/src/components/Tabs/index.tsx
@@ -1,1 +1,0 @@
-export { default as Tabs, type TabsProps } from './Tabs'
